del register controller and move it to the Auth controller
